+++ mod/Mods/tech/DATIS/Options/optionsData.lua
@@ -1,8 +1,7 @@
 cdata = {
     DCSATIS = _("DCS ATIS"),
     DATIS_CLOUD_ACCESS_KEY = _("Google Cloud Access Key:"),
+    DATIS_SRS_SERVER_PORT = _("SRS Server Port (5002):"),
     DATIS_DEBUG_LOGGING = _("Enable Debug Logging"),
-    DATIS_DEBUG_LOGGING_HELP = _(
-        "Add additional log output Saved Games\\Logs\\DATIS.log. This should only be enabled for investigating issues, since it otherwise increases the log file size unnecessarily."
-    )
+    DATIS_DEBUG_LOGGING_HELP = _("Add additional log output Saved Games\\Logs\\DATIS.log.")
 }