+++ mod/Mods/tech/DATIS/Options/optionsDb.lua
@@ -2,5 +2,6 @@ local DbOption = require("Options.DbOption")
 
 return {
     gcloudAccessKey = DbOption.new():setValue(""):editbox(),
+    srsPort = DbOption.new():setValue("5002"):editbox(),
     debugLoggingEnabled = DbOption.new():setValue(false):checkbox()
 }